Break up global-tabs
